+++ massiv/src/Data/Massiv/Array/Numeric.hs
@@ -63,7 +63,6 @@ module Data.Massiv.Array.Numeric
   ) where
 
 import           Data.Massiv.Array.Delayed.Internal
-import           Data.Massiv.Array.Manifest (U)
 import           Data.Massiv.Array.Manifest.Internal (compute)
 import           Data.Massiv.Array.Ops.Fold         as A
 import           Data.Massiv.Array.Ops.Map          as A