+++ massiv/src/Data/Massiv/Array/Numeric/Integral.hs
@@ -1,14 +1,44 @@
+{-# LANGUAGE BangPatterns     #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE BangPatterns #-}
-module Data.Massiv.Array.Numeric.Integral where
+module Data.Massiv.Array.Numeric.Integral
+  (
+  -- $integral_intro
+  --
+  -- * Integral Approximation
+  -- ** Midpoint Rule
+    midpointRule
+  , midpointStencil
+  -- ** Trapezoid Rule
+  , trapezoidRule
+  , trapezoidStencil
+  -- ** Simpson's Rule
+  , simpsonsRule
+  , simpsonsStencil
+  -- * General Integral approximation
+  , integrateWith
+  , integralApprox
+  -- * From functions
+  -- ** Sampled at the edge
+  , fromFunction
+  -- ** Sampled at the midpoint
+  , fromFunctionMidpoint
+  ) where
 
-import           Data.Massiv.Core.Common
+import           Data.Coerce
 import           Data.Massiv.Array.Delayed
 import           Data.Massiv.Array.Manifest.Internal
+import           Data.Massiv.Array.Ops.Transform     (extract')
 import           Data.Massiv.Array.Stencil
-import           Data.Massiv.Array.Ops.Transform (extract')
-import Data.Coerce
+import           Data.Massiv.Core.Common
 
+
+-- |
+--
+-- __Midpoint Rule__
+--
+-- \[
+-- \int_{{\,a}}^{{\,b}}{{f\left( x \right)\,dx}} \approx \Delta x \cdot \,f\left( {x_1 + \frac{\Delta x}{2}} \right) + \Delta x \cdot \,f\left( {x_2 + \frac{\Delta x}{2}} \right) +  \cdots  + \Delta x \cdot \,f\left( {x_n + \frac{\Delta x}{2}} \right)
+-- \]
 midpointStencil ::
      (Fractional e, Index ix)
   => e -- ^ Δx - distance between sample points
@@ -18,8 +48,16 @@ midpointStencil ::
 midpointStencil dx dim k =
   makeStencilDef 0 (setIndex' (pureIndex 1) dim k) zeroIndex $ \g ->
     pure dx * loop 0 (< k) (+ 1) 0 (\i -> (+ g (setIndex' zeroIndex dim i)))
+{-# INLINE midpointStencil #-}
 
 
+-- |
+--
+-- __Trapezoid Rule__
+--
+-- \[
+-- \int_{{\,a}}^{{\,b}}{{f\left( x \right)\,dx}} \approx \frac{{\Delta x}}{2}\cdot\left( {f\left( {{x_0}} \right) + f\left( {{x_1}} \right)} \right) + \frac{{\Delta x}}{2}\cdot\left( {f\left( {{x_1}} \right) + f\left( {{x_2}} \right)} \right) +  \cdots  + \frac{{\Delta x}}{2}\cdot\left( {f\left( {{x_{n - 1}}} \right) + f\left( {{x_n}} \right)} \right)
+-- \]
 trapezoidStencil ::
      (Fractional e, Index ix)
   => e -- ^ Δx - distance between sample points
@@ -31,8 +69,16 @@ trapezoidStencil dx dim k =
     pure dx / 2 *
     (loop 1 (< k) (+ 1) (g zeroIndex) (\i -> (+ 2 * g (setIndex' zeroIndex dim i))) +
      g (setIndex' zeroIndex dim k))
+{-# INLINE trapezoidStencil #-}
 
 
+-- |
+--
+-- __Simpson's Rule__
+--
+-- \[
+-- \int_{{\,a}}^{{\,b}}{{f\left( x \right)\,dx}} \approx \frac{{\Delta x}}{3}\cdot\left( {f\left( {{x_0}} \right) + 4\cdot f\left( {{x_1}} \right) + f\left( {{x_2}} \right)} \right) + \frac{{\Delta x}}{3}\cdot\left( {f\left( {{x_2}} \right) + 4\cdot f\left( {{x_3}} \right) + f\left( {{x_4}} \right)} \right) +  \cdots  + \frac{{\Delta x}}{3}\cdot\left( {f\left( {{x_{n - 2}}} \right) + 4\cdot f\left( {{x_{n - 1}}} \right) + f\left( {{x_n}} \right)} \right)
+-- \]
 simpsonsStencil ::
      (Fractional e, Index ix)
   => e -- ^ Δx - distance between sample points
@@ -46,8 +92,9 @@ simpsonsStencil dx dim k =
               !newAcc = acc + prev + 4 * g (setIndex' zeroIndex dim (i + 1)) + fx3
            in (fx3, newAcc)
      in pure dx / 3 * snd (loop 2 (< k - 1) (+ 2) (simAcc 0 (g zeroIndex, 0)) simAcc)
+{-# INLINE simpsonsStencil #-}
 
-
+-- | Integrate with a stencil along a particular dimension.
 integrateWith ::
      (Fractional e, Load DW ix e, Mutable r ix e)
   => (Dim -> Int -> Stencil ix e e)
@@ -56,170 +103,102 @@ integrateWith ::
   -> Array r ix e
   -> Array r ix e
 integrateWith stencil dim n arr =
-  computeWithStride (Stride nsz) $
-  mapStencil (Fill 0) (stencil dim n) arr
+  computeWithStride (Stride nsz) $ mapStencil (Fill 0) (stencil dim n) arr
   where
-    nsz = setIndex' (pureIndex 1) dim n
+    !nsz = setIndex' (pureIndex 1) dim n
+{-# INLINE integrateWith #-}
 
 
-
-integralApprox1 ::
-     (Fractional e, Mutable r Ix1 e)
-  => (e -> Dim -> Int -> Stencil Ix1 e e) -- ^ Integration Stencil
-  -> Ix1 -- ^ Result size of the matrix
-  -> e -- ^ Length of interval @d@
-  -> Array r Ix1 e
-  -> Array M Ix1 e
-integralApprox1 stencil sz d arr = extract' 0 sz $ toManifest $ applyStencil 1 (getIndex' n 1) arr
-  where
-    applyStencil = integrateWith (stencil dx)
-    n = liftIndex2 div (size arr) sz
-    nFrac = fromIntegral $ getIndex' n 1
-    dx = d / nFrac
-
-integralApprox2 ::
-     (Fractional e, Mutable r Ix2 e)
-  => (e -> Dim -> Int -> Stencil Ix2 e e) -- ^ Integration Stencil
-  -> Ix2 -- ^ Result size of the matrix
-  -> e -- ^ Length of interval @d@
-  -> Array r Ix2 e
-  -> Array M Ix2 e
-integralApprox2 stencil sz d arr =
-  extract' 0 sz $
-  toManifest $
-  applyStencil 2 (getIndex' n 2) $
-  applyStencil 1 (getIndex' n 1) arr
-  where
-    applyStencil = integrateWith (stencil dx)
-    n = liftIndex2 div (size arr) sz
-    nFrac = fromIntegral $ getIndex' n 1
-    dx = d / nFrac
-
-
--- | Compute an approximation of integral using supplied rule in form of a `Stencil`.
+-- | Compute an approximation of integral using a supplied rule in a form of `Stencil`.
 integralApprox ::
      (Fractional e, Load DW ix e, Mutable r ix e)
   => (e -> Dim -> Int -> Stencil ix e e) -- ^ Integration Stencil
-  -> ix -- ^ Result size of the matrix
-  -> Int
-  -> e -- ^ Length of interval @d@
-  -> Array r ix e
+  -> e -- ^ @d@ - Length of interval per cell
+  -> ix -- ^ @sz@ - Result size of the matrix
+  -> Int -- ^ @n@ - number of samples
+  -> Array r ix e -- ^ Array with values of @f(x,y,..)@ that will be used as source for integration.
   -> Array M ix e
-integralApprox stencil sz n d arr =
+integralApprox stencil d sz n arr =
   extract' zeroIndex sz $ toManifest $ loop 1 (<= coerce (dimensions sz)) (+ 1) arr applyStencil
   where
     !dx = d / fromIntegral n
     applyStencil dim = integrateWith (stencil dx) (Dim dim) n
+    {-# INLINE applyStencil #-}
+{-# INLINE integralApprox #-}
 
 
-midpointRule1 ::
-     (Fractional e, Mutable r Ix1 e)
-  => r -- ^ Intermediate array representation
-  -> (e -> e) -- ^ f(x) - function to integrate
-  -> e -- ^ a - startig point
-  -> e -- ^ d - distance per matrix cell
-  -> Ix1 -- ^ end vector size
-  -> Int -- ^ number of sample points per cell
-  -> Array M Ix1 e
-midpointRule1 r f a d sz n = integralApprox midpointStencil sz n d $ computeAs r arr
-  where arr = fromFunctionMidpoint fromFunction1 f a d sz n
+-- | Use midpoint rule to approximate an integral.
+midpointRule ::
+     (Fractional e, Load DW ix e, Mutable r ix e)
+  => Comp -- ^ Computation strategy
+  -> r -- ^ Intermediate array representation
+  -> ((Int -> e) -> ix -> e) -- ^ @f(x,y,...)@ - function to integrate
+  -> e -- ^ @a@ - startig point
+  -> e -- ^ @d@ - distance per matrix cell
+  -> ix -- ^ @sz@ - end matrix size
+  -> Int -- ^ @n@ - number of sample points per cell in each direction
+  -> Array M ix e
+midpointRule comp r f a d sz n =
+  integralApprox midpointStencil d sz n $ computeAs r $ fromFunctionMidpoint comp f a d sz n
+{-# INLINE midpointRule #-}
 
 
-midpointRule2 ::
-     (Fractional e, Mutable r Ix2 e)
-  => r -- ^ Intermediate array representation
-  -> (e -> e -> e) -- ^ f(x, y) - function to integrate
-  -> e -- ^ a - startig point
-  -> e -- ^ d - distance per matrix cell
-  -> Ix2 -- ^ end matrix size
-  -> Int -- ^ number of sample points per cell in each direction
-  -> Array M Ix2 e
-midpointRule2 r f a d sz n = integralApprox midpointStencil sz n d $ computeAs r arr
-  where arr = fromFunctionMidpoint fromFunction2 f a d sz n
+-- | Use trapezoid rule to approximate an integral.
+trapezoidRule ::
+     (Fractional e, Load DW ix e, Mutable r ix e)
+  => Comp -- ^ Computation strategy
+  -> r -- ^ Intermediate array representation
+  -> ((Int -> e) -> ix -> e) -- ^ @f(x,y,...)@ - function to integrate
+  -> e -- ^ @a@ - startig point
+  -> e -- ^ @d@ - distance per matrix cell
+  -> ix -- ^ @sz@ - end matrix size
+  -> Int -- ^ @n@ - number of sample points per cell in each direction
+  -> Array M ix e
+trapezoidRule comp r f a d sz n =
+  integralApprox trapezoidStencil d sz n $ computeAs r $ fromFunction comp f a d sz n
+{-# INLINE trapezoidRule #-}
+
+-- | Use Simpson's rule to approximate an integral.
+simpsonsRule ::
+     (Fractional e, Load DW ix e, Mutable r ix e)
+  => Comp -- ^ Computation strategy
+  -> r -- ^ Intermediate array representation
+  -> ((Int -> e) -> ix -> e) -- ^ @f(x,y,...)@ - function to integrate
+  -> e -- ^ @a@ - startig point
+  -> e -- ^ @d@ - distance per matrix cell
+  -> ix -- ^ @sz@ - end matrix size
+  -> Int -- ^ @n@ - number of sample points per cell in each direction
+  -> Array M ix e
+simpsonsRule comp r f a d sz n =
+  integralApprox trapezoidStencil d sz n $ computeAs r $ fromFunction comp f a d sz n
+{-# INLINE simpsonsRule #-}
 
 
-trapezoidRule2 ::
-     (Fractional e, Mutable r Ix2 e)
-  => r -- ^ Intermediate array representation
-  -> (e -> e -> e) -- ^ f(x, y) - function to integrate
-  -> e -- ^ a - startig point
-  -> e -- ^ d - distance per matrix cell
-  -> Ix2 -- ^ end matrix size
-  -> Int -- ^ number of sample points per cell in each direction
-  -> Array M Ix2 e
-trapezoidRule2 r f a d sz n = integralApprox trapezoidStencil sz n d $ computeAs r arr
-  where arr = fromFunction2 f a d sz n
-
-
-simpsonsRule2 ::
-     (Fractional e, Mutable r Ix2 e)
-  => r -- ^ Intermediate array representation
-  -> (e -> e -> e) -- ^ f(x, y) - function to integrate
-  -> e -- ^ a - startig point
-  -> e -- ^ d - distance per matrix cell
-  -> Ix2 -- ^ end matrix size
-  -> Int -- ^ number of sample points per cell in each direction
-  -> Array M Ix2 e
-simpsonsRule2 r f a d sz n = integralApprox simpsonsStencil sz n d $ computeAs r arr
-  where arr = fromFunction2 f a d sz n
-
-
--- integrate stencil r f a d sz n = integralApprox stencil sz n d $ computeAs r arr
---   where arr = fromFunctionMidpoint fromFunction2 f a d sz n
-
-
-
--- midpointRuleIx fromFunIx r f a d sz n = integralApprox midpointStencil sz d $ computeAs r arr
---   where arr = fromFunctionMidpoint fromFun f a d sz n
-
-
-fromFunction1
-  :: Fractional a =>
-     (a -> b) -> a -> a -> Ix1 -> Int -> Array D Ix1 b
-fromFunction1 f a d sz n =
-  fmap (\i -> f (scale i)) $ range' Par 0 (n * sz)
+-- | Create an array from a function with sample points at the edges
+fromFunction
+  :: (Index ix, Fractional a) =>
+     Comp -> ((Int -> a) -> ix -> e) -> a -> a -> ix -> Int -> Array D ix e
+fromFunction comp f a d sz n =
+  fmap (\ix -> f scale ix) $ range' comp zeroIndex (liftIndex (n *) sz)
   where
     nFrac = fromIntegral n
     scale i = a + d * fromIntegral i / nFrac
+    {-# INLINE scale #-}
+{-# INLINE fromFunction #-}
 
 
+-- | Create an array from a function with sample points in the middle of cells
 fromFunctionMidpoint
-  :: (Integral a, Fractional t1) =>
-     (t2 -> t1 -> t1 -> t3 -> a -> t4)
-     -> t2 -> t1 -> t1 -> t3 -> a -> t4
-fromFunctionMidpoint fromFunction f a d sz n = fromFunction f (a + dx2) d sz n
+  :: (Index ix, Fractional a) =>
+     Comp -> ((Int -> a) -> ix -> e) -> a -> a -> ix -> Int -> Array D ix e
+fromFunctionMidpoint comp f a d sz n =
+  fmap (\ix -> f scale ix) $ range' comp zeroIndex (liftIndex (\i -> n * i - 1) sz)
   where
     nFrac = fromIntegral n
     dx2 = d / nFrac / 2
-
-
--- fromFunctionMidpoint1
---   :: Fractional a =>
---      (a -> b) -> a -> a -> Ix1 -> Int -> Array D Ix1 b
--- fromFunctionMidpoint1 f a d sz n = fromFunction1 f (a + dx2) d sz n
---   where
---     nFrac = fromIntegral n
---     dx2 = d / nFrac / 2
-
-
-fromFunction2
-  :: Fractional a =>
-     (a -> a -> e) -> a -> a -> Ix2 -> Int -> Array D Ix2 e
-fromFunction2 f a d sz n =
-  fmap (\(i :. j) -> f (scale i) (scale j)) $ range' Par zeroIndex (pureIndex n * sz)
-  where
-    nFrac = fromIntegral n
-    scale i = a + d * fromIntegral i / nFrac
-
--- fromFunctionMidpoint2
---   :: Fractional a =>
---      (a -> a -> e) -> a -> a -> Ix2 -> Int -> Array D Ix2 e
--- fromFunctionMidpoint2 f a d sz n = fromFunction2 f (a + dx2) d sz n
---   where
---     nFrac = fromIntegral n
---     dx2 = d / nFrac / 2
-
-
+    scale i = dx2 + a + d * fromIntegral i / nFrac
+    {-# INLINE scale #-}
+{-# INLINE fromFunctionMidpoint #-}
 
 
 -- TODO: make this function external
@@ -229,41 +208,37 @@ range' comp ixFrom ixTo =
   makeArray comp sz (\ix -> liftIndex2 (+) ixFrom ix)
   where
     sz = liftIndex2 (-) (liftIndex (+ 1) ixTo) ixFrom
+{-# INLINE range' #-}
 
 
-gaussianIx2 :: Floating a => a -> a -> a -> a
-gaussianIx2 stdDev y x = exp (-(x ^ (2 :: Int) + y ^ (2 :: Int)) / var2) / (var2 * pi)
-  where
-    var2 = 2 * stdDev ^ (2 :: Int)
-
-ga1 :: Floating a => a -> a
-ga1 x = exp (x ^ (2 :: Int))
-
--- calculateDx :: (Fractional e, Manifest r ix e) => Dim -> Int -> Array r ix e -> e
--- calculateDx dim n arr = ((arr ! setIndex' zeroIndex dim n) - (arr ! zeroIndex)) / fromIntegral n
-
-
-
-
--- trapezoidRule ::
---      (Fractional e, Load DW ix e, Manifest r ix e, Mutable r' ix e)
---   => r'
---   -> Dim
---   -> Int
---   -> e
---   -> Array r ix e
---   -> Array r' ix e
--- trapezoidRule _ dim n dx = integrateWith (trapezoidStencil dim (n + 1) dx) dim n
-
-
--- simpsonsRule ::
---      (Fractional e, Load DW ix e, Manifest r ix e, Mutable r' ix e)
---   => r'
---   -> Dim
---   -> Int
---   -> e
---   -> Array r ix e
---   -> Array r' ix e
--- simpsonsRule _ dim n dx = integrateWith (simpsonStencil dim (n + 1) dx) dim n
-
-
+-- $integral_intro
+--
+-- The inspiration for the code in this module was taken from [Paul Dawkins Online
+-- Notes](http://tutorial.math.lamar.edu). In particular the page on [Integral
+-- Approximation](http://tutorial.math.lamar.edu/Classes/CalcII/ApproximatingDefIntegrals.aspx),
+-- so if you need to brush up on some theory it is a great place to start.
+--
+-- Implementation-wise approximation here relies heavily on stencils with stride, as such
+-- computation is fast and is automatically parallelizable.
+--
+-- Here is an example of where this can be useful.
+--
+-- Calculate a value of an integral of a function, say a gaussian @f(x) = e^(x^2)@ on interval
+-- @[0, 2]@ (as in Paul's tutorial above). For this we define a function @f@, an array with equally
+-- spaced (/dx/) sample input values and apply the function to that array, which will give us an
+-- array of @n + 1@ sample points, or looking from a different angle @n@ intervals.
+--
+-- >λ> let f x = exp ( x ^ 2 )
+-- >λ> fromFunction Seq (\ scale x -> f (scale x)) 0 2 1 4
+-- >(Array D Seq (5)
+-- >  [ 1.0,1.2840254166877414,2.718281828459045,9.487735836358526,54.598150033144236 ])
+--
+-- Once we have that array of sample points ready we could use `integralApprox` and one of the
+-- stencils to compute an integral, but there already functions that will do both steps for you:
+--
+-- >λ> trapezoidRule Seq U (\ scale x -> f (scale x)) 0 2 1 4
+-- >(Array M Seq (1)
+-- >  [ 20.644559049038715 ])
+--
+-- Scaling function @scale@ is what will change an array index into equally spaced and
+-- appropriately shifted values of @x, y, ...@ before they can get applied to @f(x, y, ...)@